数据存储写入控制。

+++ latke/src/main/java/org/b3log/latke/repository/AbstractRepository.java
@@ -42,7 +42,7 @@ import org.json.JSONObject;
  * </p>
  *
  * @author <a href="mailto:DL88250@gmail.com">Liang Ding</a>
- * @version 1.0.0.7, Feb 28, 2012
+ * @version 1.0.0.8, Apr 9, 2012
  */
 public abstract class AbstractRepository implements Repository {
 
@@ -100,14 +100,21 @@ public abstract class AbstractRepository implements Repository {
 
     @Override
     public String add(final JSONObject jsonObject) throws RepositoryException {
+        if (!isWritable()) {
+            throw new RepositoryException("The repository[name=" + getName() + "] is not writable at present");
+        }
+
         Repositories.check(getName(), jsonObject, Keys.OBJECT_ID);
 
         return repository.add(jsonObject);
     }
 
     @Override
-    public void update(final String id, final JSONObject jsonObject)
-            throws RepositoryException {
+    public void update(final String id, final JSONObject jsonObject) throws RepositoryException {
+        if (!isWritable()) {
+            throw new RepositoryException("The repository[name=" + getName() + "] is not writable at present");
+        }
+
         Repositories.check(getName(), jsonObject, Keys.OBJECT_ID);
 
         repository.update(id, jsonObject);
@@ -115,6 +122,10 @@ public abstract class AbstractRepository implements Repository {
 
     @Override
     public void remove(final String id) throws RepositoryException {
+        if (!isWritable()) {
+            throw new RepositoryException("The repository[name=" + getName() + "] is not writable at present");
+        }
+
         repository.remove(id);
     }
 
@@ -190,6 +201,16 @@ public abstract class AbstractRepository implements Repository {
     @Override
     public Cache<String, Serializable> getCache() {
         return repository.getCache();
+    }
+
+    @Override
+    public boolean isWritable() {
+        return repository.isWritable();
+    }
+
+    @Override
+    public void setWritable(final boolean writable) {
+        repository.setWritable(writable);
     }
 
     /**
